+++ jason/demo/minikube.vals.yaml
@@ -11,7 +11,7 @@ server:
     server:
       client:
         http:
-          public_url: wss://medea-demo.test
+          public_url: wss://medea-demo.test/ws
 
     turn:
       host: medea-demo.test
