+++ jason/demo/staging.vals.yaml
@@ -11,7 +11,7 @@ server:
     server:
       client:
         http:
-          public_url: wss://demo.medea.stg.t11913.org
+          public_url: wss://demo.medea.stg.t11913.org/ws
           bind_port: 9980
     turn:
       host: demo.medea.stg.t11913.org
